feat: move injection of `t` function
